Untangle other snapshots from ExportList.test.js because of importing fixture functions
